Back to all tests passing after another round of aggressive... file renaming and refactoring of the algorithms.
