+++ index.js
@@ -1,29 +1,48 @@
-// library.js - encapsule/jsgraph module exports
+// Encapsule/jsgraph/index.js
+// Public package exports for jsgraph module.
 
-var digraph = require('./src/digraph');
-var digraph_bfs = require('./src/digraph-bfs');
-var createBreadthFirstSearchContext = require('./src/digraph-bfs-context');
+var DirectedGraph = require('./src/digraph');
+
+// Breadth-first Traverse (BFT): Unifies breadth-first visit and search algorithms.
+var BFT = require('./src/digraph-algorithm-bft');
+var createBreadthFirstTraverseContext = require('./src/digraph-algorithm-bft-context');
+
+// Depth-First Traverse (DFT): Unifies depth-first visit and search algorithms.
+
 var digraph_dfs = require('./src/digraph-dfs');
-var digraph_transpose = require('./src/digraph-transpose');
+var digraph_transpose = require('./src/digraph-algorithm-transpose');
+
+// NEW STUFF :: HOLD OFF ON DFT FOR NOW...
+// var DFT = require('./src/digraph-algorithm-dft');
+// var createDepthFirstTraverseContext = require('./src/digraph-algorithm-dft-context');
+
+var digraph_transpose = require('./src/digraph-algorithm-transpose');
 
 
 module.exports = {
 
     // Generic directed graph container.
-    DirectedGraph: digraph, 
-    // Directed graph-specific algorithms and transforms.
+    DirectedGraph: DirectedGraph, 
+
+    // Directed graph algorithms and transforms.
     directed: {
-        // Generic, in-memory directed graph container object.
+
         // Directed graph transposition algorithm.
         transpose: digraph_transpose,
-        // Directed graph breadth-first visit and search algorithms.
-        createBreadthFirstSearchContext: createBreadthFirstSearchContext,
-        breadthFirstVisit: digraph_bfs.breadthFirstSearch,
-        breadthFirstSearch: digraph_bfs.breadthFirstSearch,
+
+        // Directed graph breadth-first visit and search algorithms (unified API).
+        breadthFirstTraverse: BFT,
+        createBreadthFirstTraverseContext: createBreadthFirstTraverseContext,
+
         // Directed graph depth-first visit and search algorithms.
         createDepthFirstSearchContext: digraph_dfs.createDepthFirstSearchContext,
         depthFirstVisit: digraph_dfs.depthFirstVisit,
-        depthFirstSearch: digraph_dfs.depthFirstSearch
+        depthFirstSearch: digraph_dfs.depthFirstSearch        
+
+        // NEW STUFF :: 
+        // depthFirstTraverse: DTF,
+        // createDepthFirstSearchContext: createDepthFirstTraverseContext
+
     },
 
     // Someday this too will be available.
