+++ src/digraph-algorithm-bft-colors.js
