+++ src/digraph-algorithm-bft-request.js
@@ -2,33 +2,30 @@
 //
 
 var helperFunctions = require('./helper-functions');
-var createBreadthFirstSearchContext = require('./digraph-bfs-context');
+var createBreadthFirstTraverseContext = require('./digraph-algorithm-bft-context');
 
 /*
-
   request = {
-  digraph: reference to jsgraph.DirectedGraph container object (required)
-  visitor: reference to jsgraph BFV visitor object (required)
-  options: {
-  startVector: reference to a vertex ID string, or an array of vertex ID strings (optional)
-  Note: if ommitted, BFS uses the digraph's root vertex set as the start vertex set
-  signalStart: Boolean flag (optional - default is true if ommitted)
-  Note: By default, BFS will call startVertex on each search root vertex.
-  In advanced scenarios you may wish to override this behavior.
-  searchContext: reference to BFS search context object (optional)
-  Note: By default, BFS allocates the search context internally and returns it to
-  the caller. In advanced scenarios you may wish to provide a pre-initialized
-  (or potentially pre-colored) search context object.
-  }
+      digraph: reference to jsgraph.DirectedGraph container object (required)
+      visitor: reference to jsgraph BFV visitor object (required)
+      options: {
+          startVector: reference to a vertex ID string, or an array of vertex ID strings (optional)
+              Note: if ommitted, BFS uses the digraph's root vertex set as the start vertex set
+          signalStart: Boolean flag (optional - default is true if ommitted)
+              Note: By default, BFS will call startVertex on each search root vertex.
+              In advanced scenarios you may wish to override this behavior.
+          searchContext: reference to BFS search context object (optional)
+              Note: By default, BFS allocates the search context internally and returns it to
+              the caller. In advanced scenarios you may wish to provide a pre-initialized
+              (or potentially pre-colored) search context object.
+          }
+      }
   }
 
   response = {
-  error: null indicating success or a string containing an explanation of the failure
-  result: {
-  searchCompleted: Boolean flag
-  searchContext: reference to the BFS search context object
-  } // or null to indicate a failure
-
+      error: null or string explaining why result is null
+      result: BFS search context object
+  }
 */
 
 module.exports = function (request_) {
@@ -38,8 +35,8 @@ module.exports = function (request_) {
     var nrequest = null;
     var inBreakScope = false;
 
-    var createSearchContext = function() {
-        var response = createBreadthFirstSearchContext({ digraph: nrequest.digraph });
+    var createTraverseContext = function() {
+        var response = createBreadthFirstTraverseContext({ digraph: nrequest.digraph });
         var result = null;
         if (response.error) {
             errors.unshift(response.error);
@@ -125,7 +122,7 @@ module.exports = function (request_) {
         
         helperFunctions.setPropertyValueIfUndefined(nrequest.options, 'startVector', getRootVertices);
         helperFunctions.setPropertyValueIfUndefined(nrequest.options, 'signalStart', true);
-        helperFunctions.setPropertyValueIfUndefined(nrequest.options, 'searchContext', createSearchContext);
+        helperFunctions.setPropertyValueIfUndefined(nrequest.options, 'searchContext', createTraverseContext);
 
         response.result = nrequest;
