+++ src/digraph-algorithm-transpose.js
