--- test/test-digraph-algorithm-bfsv.js
@@ -1,5 +1,0 @@
-// Encapsule/jsgraph/test/test-digraph-algorithm-bsfv.js
-
-require('./test-digraph-algorithm-bfv');
-require('./test-digraph-algorithm-bfs');
-
