+++ test/test-digraph-algorithm-bft.js
@@ -1,0 +1,5 @@
+// Encapsule/jsgraph/test/test-digraph-algorithm-bsfv.js
+
+require('./test-digraph-algorithm-bft-visit');
+require('./test-digraph-algorithm-bft-search');
+
