+++ test/test-digraph-algorithms.js
@@ -1,5 +1,5 @@
 // Encapsule/jsgraph/test/test-jsgraph-algorithms.js
 
 require('./test-digraph-algorithm-transpose');
-require('./test-digraph-algorithm-bfsv');
+require('./test-digraph-algorithm-bft');
 require('./test-digraph-algorithm-dfsv');
