+++ test/test-jsgraph.js
@@ -1,6 +1,6 @@
 // Top-level Mocha test file for encapsule/jsgraph module.
 
-require('./test-digraph-methods');
+require('./test-digraph-container');
 require('./test-digraph-algorithms');
 require('./test-module-exports');
